requirements.txt
fastapi
uvicorn
pillow
pytesseract
python-multipart
gunicorn